requirements.txt
# Core
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4
tqdm==4.66.4
python-dotenv==1.0.1

# Qwen API
dashscope==1.20.6

# NLP & Evaluation
pandas==2.2.2
sentence-transformers==2.5.1

# （可选）如果要做绘图/可视化表格:
# matplotlib==3.8.4

# （可选）如果以后想用 scene detection:
# scenedetect==0.6.4

# （可选）如果要把报告导出为 HTML/PDF（生成图文报告）:
# jinja2==3.1.4
# markdown==3.6